Get the Table Schema of a Data Package resource

Returns the Table Schema of a Data Package resource, i.e. the content of its schema property, describing the resource's fields, data types, relationships, and missing values. The resource must be a Tabular Data Resource.

get_schema(resource_name, package)

Arguments

resource_name
Name of the resource.

package
Package object, see read_package().

Value

List object describing the Table Schema.

Examples

# Read datapackage.json file
package <- read_package(system.file("extdata", "datapackage.json", package = "datapackage"))
#> Please make sure you have the right to access data from this Data Package for your proposed use.
#> Follow applicable norms or requirements to credit the dataset and its authors.

# Get table schema of resource "observations"
get_schema("observations", package)
#> $fields
#> $fields[[1]]
#> $fields[[1]]$name
#> [1] "observation_id"
#> 
#> $fields[[1]]$type
#> [1] "string"
#> 
#> $fields[[1]]$constraints
#> $fields[[1]]$constraints$required
#> [1] TRUE
#> 
#> $fields[[1]]$constraints$unique
#> [1] TRUE
#> 
#> 
#> 
#> $fields[[2]]
#> $fields[[2]]$name
#> [1] "deployment_id"
#> 
#> $fields[[2]]$type
#> [1] "string"
#> 
#> $fields[[2]]$constraints
#> $fields[[2]]$constraints$required
#> [1] TRUE
#> 
#> 
#> 
#> $fields[[3]]
#> $fields[[3]]$name
#> [1] "timestamp"
#> 
#> $fields[[3]]$type
#> [1] "datetime"
#> 
#> $fields[[3]]$format
#> [1] "%Y-%m-%dT%H:%M:%S%z"
#> 
#> $fields[[3]]$constraints
#> $fields[[3]]$constraints$required
#> [1] TRUE
#> 
#> 
#> 
#> $fields[[4]]
#> $fields[[4]]$name
#> [1] "scientific_name"
#> 
#> $fields[[4]]$type
#> [1] "string"
#> 
#> $fields[[4]]$constraints
#> $fields[[4]]$constraints$required
#> [1] FALSE
#> 
#> 
#> 
#> $fields[[5]]
#> $fields[[5]]$name
#> [1] "count"
#> 
#> $fields[[5]]$type
#> [1] "integer"
#> 
#> $fields[[5]]$constraints
#> $fields[[5]]$constraints$required
#> [1] FALSE
#> 
#> $fields[[5]]$constraints$minimum
#> [1] 1
#> 
#> 
#> 
#> $fields[[6]]
#> $fields[[6]]$name
#> [1] "age"
#> 
#> $fields[[6]]$type
#> [1] "string"
#> 
#> $fields[[6]]$constraints
#> $fields[[6]]$constraints$required
#> [1] FALSE
#> 
#> $fields[[6]]$constraints$enum
#> [1] "adult"     "subadult"  "juvenile"  "offspring" "undefined"
#> 
#> 
#> 
#> $fields[[7]]
#> $fields[[7]]$name
#> [1] "comments"
#> 
#> $fields[[7]]$type
#> [1] "string"
#> 
#> $fields[[7]]$constraints
#> $fields[[7]]$constraints$required
#> [1] FALSE
#> 
#> 
#> 
#> 
#> $missingValues
#> [1] ""    "NA"  "NaN"
#> 
#> $primaryKey
#> [1] "observation_id"
#> 
#> $foreignKeys
#> $foreignKeys[[1]]
#> $foreignKeys[[1]]$fields
#> [1] "deployment_id"
#> 
#> $foreignKeys[[1]]$reference
#> $foreignKeys[[1]]$reference$resource
#> [1] "deployments"
#> 
#> $foreignKeys[[1]]$reference$fields
#> [1] "deployment_id"
#> 
#> 
#> 
#>