Read data from a Tabular Data Resource into a tibble

Reads data from a Data Package resource into a tibble (a Tidyverse data frame). The resource has to meet the requirements of a Tabular Data Resource. The function is a wrapper around readr::read_delim(), passing the resource properties path, CSV dialect, column names, data types, etc. Column names are taken from the provided schema, not from the header in the CSV file(s).

read_resource(package, resource_name)

Arguments

package
Package object, see read_package().

resource_name
Name of the resource to load data from.

Value

A tibble() with the resource data.

Details

The resource properties are handled as follows:

Path

path is required. It can be a local path or URL, which must resolve. Absolute path (/) and relative parent path (../) are forbidden to avoid security vulnerabilities.

When multiple paths are provided ("path": [ "myfile1.csv", "myfile2.csv"]) then data are merged into a single data frame, in the order in which the paths are listed.

Data

Inline data is not supported.

Name

name is required. It is used to find the resource with name = resource_name.

Profile

profile is required to have the value tabular-data-resource.

File encoding

encoding is required if the resource file(s) are not encoded as UTF-8. For proper values (e.g. windows-1252), see "Preferred MIME Names" in encoding. The returned data frame will always be UTF-8.

CSV Dialect

dialect properties are required if the resource file(s) deviate from the default CSV settings (see below). Only deviating properties need to be specified, e.g. a tab delimited file without a header row needs:

"dialect": {"delimiter": "\t", "header": "false"}

These are the CSV dialect properties. Some are ignored by the function:

delimiter: default ,

lineTerminator: ignored, line endings LF, CRLF and CR are interpreted automatically by readr::read_delim().

quoteChar: default "

doubleQuote: default true

escapeChar: anything but \ is ignored and it will set doubleQuote to false as these fields are mutually exclusive. You can thus not escape with \" and "" in the same file.

nullSequence: ignored, use missingValues.

skipInitialSpace: default false

header: default true

commentChar: not set by default.

caseSensitiveHeader: ignored, header is not used for column names, see Schema.

csvddfVersion: ignored.

Schema

schema is required and must follow the Table Schema specification.

Field names are used as column headers.

Field types are used as column types when provided. Types are guessed when no type is provided or it has no equivalent in R.

Field formats (especially for date, time, datetime) are ignored.

missingValues are used to interpret as NA, with "" as default.

File compression

Resource file(s) with path ending in .gz, .bz2, .xz, or .zip are automatically decompressed using default readr::read_delim() functionality. Only .gz files can be read directly from URL paths. Only the extension in path can be used to indicate compression type, the compression property is ignored.

Ignored resource properties

title

description

format

mediatype

bytes

hash

sources

licenses

Examples

# Read datapackage.json file
package <- read_package(system.file("extdata", "datapackage.json", package = "datapackage"))

# List resource names
package$resource_names

#> [1] "deployments"  "observations"

# Read data from resource "observations"
read_resource(package, "observations")

#> # A tibble: 8 x 7
#>   observation_id   deployment_id timestamp           scientific_name count age  
#>   <chr>            <chr>         <dttm>              <chr>           <int> <chr>
#> 1 089113c5-e279-4… 1             2020-01-06 15:59:17 Capreolus capr…     1 juve…
#> 2 060ab157-3a8e-4… 1             2020-01-06 15:59:17 Capreolus capr…     1 adult
#> 3 aba72fdc-2fd8-4… 1             2020-01-06 16:35:23 Lepus europaeus     1 adult
#> 4 1c044f0c-b12b-4… 1             2020-01-06 17:04:04 Lepus europaeus     1 adult
#> 5 a2fb27ab-bb8b-4… 1             2020-01-06 19:19:54 Sus scrofa          2 unde…
#> 6 cdbb4a57-2204-4… 2             2021-01-01 01:25:06 Sus scrofa          1 unde…
#> 7 3806c18c-29c5-4… 2             2021-01-01 01:25:06 Sus scrofa          1 unde…
#> 8 aa01aa7c-10dc-4… 2             2021-01-01 04:47:30 Sus scrofa          1 unde…
#> # … with 1 more variable: comments <chr>

# The above tibble is merged from 2 files listed in `path`
package$resources[[2]]$path

#> [1] "observations_1.csv" "observations_2.csv"

# With col_names and col_types derived from the resource schema
purrr::map_chr(package$resources[[2]]$schema$fields, "name")

#> [1] "observation_id"  "deployment_id"   "timestamp"       "scientific_name"
#> [5] "count"           "age"             "comments"       

purrr::map_chr(package$resources[[2]]$schema$fields, "type")

#> [1] "string"   "string"   "datetime" "string"   "integer"  "string"   "string"