Read data from a Data Package resource

Reads data from a Data Package resource into a tibble (a Tidyverse data frame). The resource has to meet the requirements of a Tabular Data Resource. The function is a wrapper around readr::read_delim(), passing the resource properties path, CSV dialect, column names, data types, etc. Column names are taken from the provided schema, not from the header in the CSV file(s).

read_resource(descriptor, resource_name)

Arguments

descriptor
Descriptor object, see read_descriptor().

resource_name
Name of the resource to load data from.

Value

Tibble with the resource data.

Details

The resource properties are handled as follows:

Path

path is required. It can be a local path or URL, which must resolve. When multiple paths are provided ("path": [ "myfile1.csv", "myfile2.csv"]) then data are merged into a single data frame, in the order in which the paths are listed.

Data

Inline data is not supported.

Name

name is required. It is used to find the resource with name = resource_name.

Profile

profile is required to have the value tabular-data-resource.

File encoding

encoding is required if resource files are not encoded as UTF-8. For proper values, see encoding. The returned data frame will always be UTF-8.

CSV Dialect

If no dialect is provided, then resource CSV files are required to follow the CSV file requirements.

If a dialect is provided, then the properties are interpreted as described in the CSV Dialect properties, with the same defaults. Exceptions are:

escapeChar: ignored if different than \.

lineTerminator: ignored.

nullSequence: ignored.

caseSensitiveHeader: ignored.

csvddfVersion: ignored.

Schema

schema is required and must follow the Table Schema specification.

Field names are used as column headers.

Field types are used as column types when provided. Types are guessed when no type is provided or it has no equivalent in R.

Field formats (especially for date, time, datetime) are ignored.

missingValues are used to interpret as NA, with "" as default.

Ignored resource properties

title

description

format

mediatype

bytes

hash

sources

licenses

Examples

descriptor <- read_descriptor(system.file("extdata", "datapackage.json", package = "datapackage"))
read_resource(descriptor, "observations")

#> # A tibble: 8 x 7
#>   observation_id   deployment_id timestamp           scientific_name count age  
#>   <chr>            <chr>         <dttm>              <chr>           <int> <chr>
#> 1 089113c5-e279-4… 1             2020-01-06 15:59:17 Capreolus capr…     1 juve…
#> 2 060ab157-3a8e-4… 1             2020-01-06 15:59:17 Capreolus capr…     1 adult
#> 3 aba72fdc-2fd8-4… 1             2020-01-06 16:35:23 Lepus europaeus     1 adult
#> 4 1c044f0c-b12b-4… 1             2020-01-06 17:04:04 Lepus europaeus     1 adult
#> 5 a2fb27ab-bb8b-4… 1             2020-01-06 19:19:54 Sus scrofa          2 unde…
#> 6 cdbb4a57-2204-4… 2             2021-01-01 01:25:06 Sus scrofa          1 unde…
#> 7 3806c18c-29c5-4… 2             2021-01-01 01:25:06 Sus scrofa          1 unde…
#> 8 aa01aa7c-10dc-4… 2             2021-01-01 04:47:30 Sus scrofa          1 unde…
#> # … with 1 more variable: comments <chr>